requirements.txt
numpy==1.25.0
matplotlib==3.5.2
opencv-python==4.10.0.84
plyfile==1.1
pywavelets==1.3.0
scipy==1.11.4
scikit-image==0.19.2
tqdm==4.66.5
tyro
comet_ml
wandb
jaxtyping
tensorboard
viser==0.1.21
appdirs
rawpy==0.19.0
pyquaternion
nuscenes-devkit
nerfacc
splines
git+https://github.com/facebookresearch/segment-anything.git
open-clip-torch
mediapy
tensorboard
plyfile